requirements.txt
accelerate>=0.30.0
diffusers>=0.30.0
huggingface-hub>=0.23.0
imageio[ffmpeg]>=2.34.0
Pillow>=10.3.0
python-dotenv>=1.0.1
PyYAML>=6.0.1
safetensors>=0.4.3
transformers>=4.41.0
